Made arrows that annotate the graph lighter and shorter
When looking at the burndown from a distance, a black vertical arrow may
look like a drop in the burndown because its tip is connected to the end
point of the burndown. To avoid this confusion the arrow color was
changed to light gray and the arrows were shortened so as to leave some
space between them and the lines they point to.